fix(search): routed dispatch keys on verbatim shape, not timeline shape
The bigger-n legs falsified the timeline criterion the same day it was
written. Three-block fused-capped vs shape_conditioned (pooled n=239):

  SSA +7.1pp | SSU -10.0pp | TR -8.3pp | pooled -5.0pp (p=0.22)

SSU questions are not timeline-shaped, so the timeline dispatch would
have sent them to fused and inherited the -10pp. The only class where
the fused leg measured positive is verbatim-shaped (assistant-content)
asks — exactly detectVerbatimShape. Dispatch is now: verbatim-shaped →
fused, everything else → shape_conditioned. VERBATIM_PATTERNS moved to
the search layer with the dispatch (the router re-exports it);
the timeline lexicon reverted to its pre-routing home and scope.
